Add Dao tests

+++ src/main/java/ua/kharkiv/nosarev/service/api/UserService.java
@@ -20,6 +20,6 @@ public interface UserService {
 
     List<User> getAllUsersByRole(UserRole role);
 
-    int getAmountOfUsers();
+    long getAmountOfUsers();
 
 }
